+++ src/HomePage.js
@@ -1,7 +1,7 @@
 // HomepPage.js
-import Header from './Components/Header';
-import Main from './Components/Main';
-import Footer from './Components/Footer';
+import Header from './components/Header';
+import Main from './components/Main';
+import Footer from './components/Footer';
 
 function HomePage() {
     return (
